Treating Atom XML content with no type as text
RFC 4287 Section 4.1.3.1 says that content elements with no type attribute should be treated as text.

The AtomParser.ReadSyndicationContent() method would always parse and add XML contents as Fields, unless there is a type attribute specifying that the content is XML. Adding a check for Atom content elements with no type attribute where the XML contents are copied to the Value property (which gets correctly copied to an IAtomEntry.Description).